+++ lua/base/predict.lua
@@ -1,3 +1,0 @@
-//DAK loader Predict
-
-Script.Load("lua/base/shared.lua")