+++ lua/configs/guimenubase.lua
@@ -1,9 +1,0 @@
-//GUIMenuBase config
-
-local function SetupDefaultConfig()
-	local DefaultConfig = { }
-	DefaultConfig.kMenuUpdateRate = 2
-	return DefaultConfig
-end
-
-DAK:RegisterEventHook("PluginDefaultConfigs", {PluginName = "guimenubase", DefaultConfig = SetupDefaultConfig })